demo: move panelreel demo to its own file
